requirements.txt
# Web Framework
flask==2.3.3
werkzeug==2.3.7
flask-cors==4.0.0
flask-limiter==3.5.0
gunicorn==21.2.0

# Database and Caching
redis==5.0.1

# Data Validation and Schema
pydantic==2.5.2
marshmallow==3.20.1

# HTTP & Networking
requests==2.31.0
tenacity==8.2.3
httpx==0.25.1

# LLM Frameworks
openai==1.5.0
langchain-core>=0.1.42,<0.2.0
langchain>=0.0.350
langgraph==0.0.38
langchain-openai>=0.0.2
langchain-community>=0.0.11

# Configuration & Environment
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1

# Utilities
python-dateutil==2.8.2
loguru==0.7.2

# Type Hints
typing-extensions>=4.7.0
